+++ CHANGELOG.md
@@ -1,17 +1,201 @@
 # Changelog
 
-## 119.0.6045.192
+## 115.6.0-44 (2023/12/25)
 
-* Supermium 119.0.6045 (2023/12/05)
-
-## 118.0.5991.0
-
-* Supermium 118.0.5991 (2023/08/24)
+* Thunderbird 115.6.0
 * Portapps 3.9.0
+* GnuPG 2.4.3
 
-## 117.0.5874.0 (2023/06/xx)
+## 91.7.0-43 (2022/04/09)
 
-* Supermium 117.0.5874
-* Portapps 3.8.0
+* Thunderbird 91.7.0
+* Portapps 3.5.0
+* Bump github.com/pierrec/lz4/v3 from 3.3.2 to 3.3.4 (#46)
 
-All credits goes to win32ss for making Supermium browser!
+## 91.1.1-42 (2021/09/19)
+
+* Thunderbird 91.1.1
+* Portapps 3.3.1
+
+## 78.10.2-41 (2021/05/22)
+
+* Thunderbird 78.10.2
+
+## 78.8.1-40 (2021/03/16)
+
+* Thunderbird 78.8.1
+* Portapps 3.3.0
+* GnuPG 2.2.27
+* No more win32 release
+
+## 78.6.0-39 (2021/01/10)
+
+* Thunderbird 78.6.0
+* Portapps 3.1.0
+* GnuPG 2.2.26
+
+## 78.4.3-38 (2020/11/13)
+
+* Thunderbird 78.4.3
+
+## 78.3.2-37 (2020/10/09)
+
+* Thunderbird 78.3.2
+
+## 78.3.0-36 (2020/09/24)
+
+* Thunderbird 78.3.0
+
+## 78.2.1-35 (2020/08/30)
+
+* Thunderbird 78.2.1
+
+## 68.12.0-34 (2020/08/30)
+
+* Thunderbird 68.12.0
+* Portapps 2.6.0
+
+## 78.1.1-33 (2020/08/11)
+
+* Thunderbird 78.1.1
+* Portapps 2.5.0
+
+## 78.0.1-32 (2020/07/28)
+
+* Thunderbird 78.0.1
+* GnuPG 2.2.21
+
+## 68.11.0-31 (2020/07/28)
+
+* Thunderbird 68.11.0
+
+## 68.10.0-30 (2020/07/12)
+
+* Thunderbird 68.10.0
+
+## 68.9.0-29 (2020/06/13)
+
+* Thunderbird 68.9.0
+
+## 68.8.0-28 (2020/05/29)
+
+* Fix addons startup (portapps/phyrox-portable#3) 
+* Portapps 2.4.4
+
+## 68.8.0-27 (2020/05/12)
+
+* Thunderbird 68.8.0
+* Allow custom `policies.json`
+* Portapps 2.2.4
+
+## 68.7.0-26 (2020/04/13)
+
+* Thunderbird 68.7.0
+* Portapps 2.1.2
+
+## 68.6.0-25 (2020/03/16)
+
+* Thunderbird 68.6.0
+
+## 68.5.0-24 (2020/02/16)
+
+* Thunderbird 68.5.0
+
+## 68.4.2-23 (2020/01/27)
+
+* Thunderbird 68.4.2
+
+## 68.4.1-22 (2020/01/20)
+
+* Thunderbird 68.4.1
+
+## 68.3.1-21 (2019/12/17)
+
+* Thunderbird 68.3.1
+* GnuPG 2.2.19
+* Add `cleanup` config
+* Portapps 1.31.0
+
+## 68.3.0-20 (2019/12/05)
+
+* Thunderbird 68.3.0
+* GnuPG 2.2.18
+
+## 68.2.2-19 (2019/12/01)
+
+* Thunderbird 68.2.2
+
+## 60.9.1-18 (2019/12/01)
+
+* Thunderbird 60.9.1
+
+## 68.2.1-17 (2019/12/01)
+
+* Following a trademark violation report (#4), name and logo have been changed
+* Portapps 1.30.1
+
+## 68.2.1-16 (2019/11/03)
+
+* Thunderbird 68.2.1
+
+## 68.2.0-15 (2019/10/23)
+
+* Thunderbird 68.2.0
+
+## 68.1.2-14 (2019/10/11)
+
+* Thunderbird 68.1.2
+* Portapps 1.28.0
+
+## 68.1.1-13 (2019/09/27)
+
+* Thunderbird 68.1.1
+
+## 68.1.0-12 (2019/09/11)
+
+* Thunderbird 68.1.0
+* Fix addons disabled on startup (#2)
+* Portapps 1.26.1
+
+## 60.9.0-11 (2019/09/11)
+
+* Thunderbird 60.9.0
+* Fix addons disabled on startup (#2)
+* Portapps 1.26.1
+
+## 68.0-8 (2019/08/29)
+
+* Thunderbird 68.0
+* Portapps 1.26.0
+
+## 60.8.0-7 (2019/07/09)
+
+* Thunderbird 60.8.0
+* GnuPG 2.2.17
+
+## 60.7.2-6 (2019/06/21)
+
+* Thunderbird 60.7.2
+
+## 60.7.1-5 (2019/06/13)
+
+* Thunderbird 60.7.1
+
+## 60.7.0-4 (2019/06/06)
+
+* Extensions break on path change (portapps/firefox-portable#2)
+
+## 60.7.0-3 (2019/05/18)
+
+* Thunderbird 60.7.0
+* Portapps 1.24.1
+
+## 60.6.1-2 (2019/05/08)
+
+* Add config param to change locale
+* Make Thunderbird multilingual
+
+## 60.6.1-1 (2019/05/06)
+
+* Initial version based on Thunderbird 60.6.1
+* GnuPG 2.2.15 embedded